requirements.txt
seleniumselenium
python-dotenv
twilio